feat: create `vite-null-export` package
