+++ packages/vite-null-export/README.md
@@ -1,0 +1,8 @@
+# vite-null-export
+
+Vite port of remix's server/client-only file convention plugin https://github.com/remix-run/remix/blob/80c6842f547b7e83b58f1963894b07ad18c2dfe2/packages/remix-dev/compiler/plugins/emptyModules.ts#L10
+
+```sh
+pnpm build
+pnpm release
+```
